requirements.txt
pyserial==3.4
requests==2.9.1
